# backend/docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
  dto.MessageResponse:
    properties:
      message:
        example: success message
        type: string
    type: object
...
  dto.ProductListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.ProductResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  dto.ProductResponse:
    description: Product information returned by the API
    properties:
      accords:
        example:
        - '["woody"'
        - '"citrus"]'
        items:
          type: string
        type: array
      brand:
        example: Dior
        type: string
      can_review:
        type: boolean
      cannot_review_reason:
        type: string
      category:
        example: Eau de Parfum
        type: string
      created_at:
        example: "2025-09-28T10:00:00Z"
        type: string
      description:
        example: A fresh and woody fragrance
        type: string
      gender:
        example: unisex
        type: string
      id:
        example: 1
        type: integer
      image_url:
        example: https://example.com/image.jpg
        type: string
      intensity:
        example: moderate
        type: string
      name:
        example: Sauvage
        type: string
      notes_base:
        example:
        - '["ambroxan"]'
        items:
          type: string
        type: array
      notes_heart:
        example:
        - '["lavender"]'
        items:
          type: string
        type: array
      notes_top:
        example:
        - '["bergamot"]'
        items:
          type: string
        type: array
      occasions:
        example:
        - '["work"'
        - '"night out"]'
        items:
          type: string
        type: array
      price:
        example: 299.99
        type: number
      price_range:
        example: premium
        type: string
      seasons:
        example:
        - '["summer"'
        - '"spring"]'
        items:
          type: string
        type: array
      slug:
        example: dior-sauvage
        type: string
      stock_quantity:
        example: 50
        type: integer
      thumbnail_url:
        example: https://example.com/image_thumb.jpg
        type: string
      updated_at:
        example: "2025-09-28T10:00:00Z"
        type: string
      weight:
        example: 100
        type: number
    type: object
...
  dto.UpdateProductRequest:
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://github.com/leoferamos/aroma-sense
  description: REST API for user Aroma Sense Eccomerce
  title: Aroma Sense API
  version: "1.0"
paths:
...
  /admin/products:
    post:
      consumes:
      - multipart/form-data
      description: Creates a new product with image upload (Admin only)
      parameters:
      - description: Product name
        in: formData
        name: name
        required: true
        type: string
      - description: Product brand
        in: formData
        name: brand
        required: true
        type: string
      - description: Product weight in ml
        in: formData
        name: weight
        required: true
        type: number
      - description: Product description
        in: formData
        name: description
        type: string
      - description: Product price
        in: formData
        name: price
        required: true
        type: number
      - description: Product category
        in: formData
        name: category
        required: true
        type: string
      - description: Product notes (fragrance notes)
        in: formData
        name: notes
        required: true
        type: array
      - description: Stock quantity
        in: formData
        name: stock_quantity
        required: true
        type: integer
      - description: Product image
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Product created successfully
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Invalid request data or missing image
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new product
      tags:
      - admin
  /admin/products/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an existing product (Admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted successfully
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - admin
    get:
      consumes:
      - application/json
      description: Retrieves a specific product by its ID (Admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get product by ID
      tags:
      - admin
    patch:
      consumes:
      - application/json
      description: Updates an existing product (Admin only)
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product update data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated successfully
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Invalid product ID or request data
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden - Admin only
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - admin
...
  /products:
    get:
      consumes:
      - application/json
      description: If `query` is provided, returns a paginated search envelope; otherwise
        returns the latest products.
      parameters:
      - description: Search term (websearch syntax)
        in: query
        name: query
        type: string
      - default: 1
        description: Page number (1-based)
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page (default 10, max 100)
        in: query
        name: limit
        type: integer
      - default: relevance
        description: 'Sort order: relevance|latest'
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Search results envelope (when query is present) or paginated
            latest (when query absent and page>1)
          schema:
            $ref: '#/definitions/dto.ProductListResponse'
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: List or search products
      tags:
      - products
  /products/{id}:
    get:
      consumes:
      - application/json
      description: Retrieves detailed information about a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            $ref: '#/definitions/dto.ProductResponse'
        "400":
          description: Invalid product ID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get product by ID
      tags:
      - products
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"